+++ build.gradle
@@ -54,6 +54,9 @@ dependencies {
     //OAuth2
     implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
 
+    //Postgre SQL
+    implementation 'org.postgresql:postgresql'
+
 }
 
 tasks.named('test') {
